requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
discord.py @ git+https://github.com/Rapptz/discord.py@45d498c1b76deaf3b394d17ccf56112fa691d160
idna==3.3
multidict==6.0.2
python-dotenv==0.19.2
requests==2.27.1
slash-util==1.3.3.1
typing_extensions==4.1.1
urllib3==1.26.8
wincertstore==0.2
yarl==1.7.2
